+++ akshare/__init__.py
@@ -2338,9 +2338,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.20 fix: fix fund_financial_fund_info_em interface
 1.9.21 fix: fix fund_financial_fund_info_em interface
 1.9.22 fix: fix fund_hk_fund_hist_em interface
+1.9.23 fix: fix bond_cb_redeem_jsl interface
 """
 
-__version__ = "1.9.22"
+__version__ = "1.9.23"
 __author__ = "AKFamily"
 
 import sys