+++ docs/changelog.md
@@ -50,6 +50,10 @@
 
 ## 更新说明
 
+1.9.23 fix: fix bond_cb_redeem_jsl interface
+
+    1. 修复 bond_cb_redeem_jsl 接口
+
 1.9.22 fix: fix fund_hk_fund_hist_em interface
 
     1. 修复 fund_hk_fund_hist_em 接口
@@ -2041,6 +2045,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.23 fix: fix bond_cb_redeem_jsl interface
 
 1.9.22 fix: fix fund_hk_fund_hist_em interface